changed from a background script to a content_script
In Chrome Extensions, background scripts live in the background they don't directly interact with pages, but instead they can be provided information by other parts of the extension and process it, uninterupted by pesky problems like windows closing.

You want to use a content script. Content scripts sit on top of a Chrome Page. They cannot access any of the Javascript already on the page (nor can the Javascript on the page mess with them) but they can access the DOM, which is what you want.

For more see the Chrome Extension docs on the subject:
https://developer.chrome.com/extensions/content_scripts